added backproject function, moved some functions to vision common utils
